+++ tanjun/dependencies/__init__.py
@@ -58,6 +58,8 @@ __all__: list[str] = [
     "SfChannelBound",
     "SfGuildBound",
     "SingleStoreCache",
+    "add_concurrency_limit",
+    "add_cooldown",
     "async_cache",
     "cached_inject",
     "callbacks",
@@ -99,6 +101,8 @@ from .limiters import ConcurrencyPreExecution
 from .limiters import CooldownPreExecution
 from .limiters import InMemoryConcurrencyLimiter
 from .limiters import InMemoryCooldownManager
+from .limiters import add_concurrency_limit
+from .limiters import add_cooldown
 from .limiters import with_concurrency_limit
 from .limiters import with_cooldown
 from .locales import AbstractLocaliser