+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [dependencies.add_cooldown][tanjun.dependencies.limiters.add_cooldown] and
+  [dependencies.add_concurrency_limit][tanjun.dependencies.limiters.add_concurrency_limit]
+  functions which act as non-decorator equivalents of the relevant `"with_`
+  functions.
+- [AbstractCooldownManager.acquire][tanjun.dependencies.limiters.AbstractCooldownManager.acquire]
+  method which returns an async context manager which acquires and releases a
+  cooldown lock for you.
+- [AbstractConcurrencyLimiter.acquire][tanjun.dependencies.limiters.AbstractConcurrencyLimiter.acquire]
+  method which returns an async context manager which acquires and releases a
+  concurrency lock for you.
+
 ## [2.14.0] - 2023-04-24
 ### Added
 - Support for declaring command options through default descriptors. These